reorganizing to add more transform classes
